+++ sync-for-commerce/src/sdk/advancedcontrols.ts
@@ -8,6 +8,7 @@ import * as operations from "./models/operations";
 import * as shared from "./models/shared";
 import { SDKConfiguration } from "./sdk";
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
+import jp from "jsonpath";
 
 /**
  * Advanced company management and sync preferences.
@@ -306,6 +307,20 @@ export class AdvancedControls {
                 break;
         }
 
+        res.next = async (): Promise<operations.ListCompaniesResponse | null> => {
+            const nextCursor = jp.value(JSON.parse(decodedRes), "");
+            if (nextCursor === undefined) {
+                return null;
+            }
+
+            return await this.listCompanies(
+                {
+                    ...req,
+                },
+                retries,
+                config
+            );
+        };
         return res;
     }