+++ sync-for-commerce/src/sdk/connections.ts
@@ -8,6 +8,7 @@ import * as operations from "./models/operations";
 import * as shared from "./models/shared";
 import { SDKConfiguration } from "./sdk";
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
+import jp from "jsonpath";
 
 export class Connections {
     private sdkConfiguration: SDKConfiguration;
@@ -338,6 +339,20 @@ export class Connections {
                 break;
         }
 
+        res.next = async (): Promise<operations.ListConnectionsResponse | null> => {
+            const nextCursor = jp.value(JSON.parse(decodedRes), "");
+            if (nextCursor === undefined) {
+                return null;
+            }
+
+            return await this.list(
+                {
+                    ...req,
+                },
+                retries,
+                config
+            );
+        };
         return res;
     }