+++ sync-for-commerce/src/sdk/models/operations/index.ts
@@ -9,6 +9,7 @@ export * from "./getconfiguration";
 export * from "./getintegrationbranding";
 export * from "./getlastsuccessfulsync";
 export * from "./getlatestsync";
+export * from "./getsyncbyid";
 export * from "./getsyncflowurl";
 export * from "./getsyncstatus";
 export * from "./getvisibleaccounts";